Bug fix: Handle selection state with respect to tools properly.
Specifically when creating, editing or selecting links modules should be deselected at the start of the change not at the end (the latter is the current, confusing, indirect behaviour).

+++ src/com/modsim/gui/view/ViewUtil.java
@@ -277,6 +277,8 @@ public class ViewUtil implements MouseListener, MouseMotionListener, MouseWheelL
                 else {
                     Port p = screenSpace_portAt(e.getX(), e.getY());
 
+                    Main.selection.clear();
+
                     //Link behaviour
                     if (p != null) {
                         tool = new MakeLinkTool();
@@ -408,7 +410,8 @@ public class ViewUtil implements MouseListener, MouseMotionListener, MouseWheelL
             }
         }
         // Delete selection
-        else if (e.getKeyCode() == KeyEvent.VK_DELETE) {
+        else if (e.getKeyCode() == KeyEvent.VK_DELETE
+                 && !Main.selection.isEmpty()) {
             Main.selection.deleteAll();
         }
         // Pass to tool
